# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SM Backend API
  version: 1.0.0
  description: API spec generated from current source controllers.
servers:
  - url: http://localhost:3000
    description: Local NestJS
  - url: https://be-sm.codekhongngu.workers.dev
    description: Cloudflare Worker
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Get users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      tags: [Users]
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
  /users/{id}:
    put:
      tags: [Users]
      summary: Replace user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
    patch:
      tags: [Users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /users/{id}/reset-password:
    patch:
      tags: [Users]
      summary: Reset user password
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                type: object
  /users/me/change-password:
    patch:
      tags: [Users]
      summary: Change own password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
  /users/units:
    get:
      tags: [Users]
      summary: Get units
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unit list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      tags: [Users]
      summary: Create unit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitRequest"
      responses:
        "201":
          description: Unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /users/units/{id}:
    patch:
      tags: [Users]
      summary: Update unit
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Unit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      tags: [Users]
      summary: Delete unit
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
  /users/{id}/role:
    patch:
      tags: [Users]
      summary: Update user role
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [EMPLOYEE, MANAGER, ADMIN]
              required: [role]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /users/{id}/unit:
    patch:
      tags: [Users]
      summary: Update user unit
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                unitId:
                  type: string
              required: [unitId]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /users/import-excel:
    post:
      tags: [Users]
      summary: Import users from Excel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Imported
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      required: [oldPassword, newPassword]
    ResetPasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
      required: [newPassword]
    CreateUnitRequest:
      type: object
      properties:
        code:
          type: string
          example: S-QUYNHON
        name:
          type: string
          example: Phòng bán hàng Quy Nhơn
        telegramGroupChatId:
          type: string
          nullable: true
        parentUnitId:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [code, name]
    Unit:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        telegramGroupChatId:
          type: string
          nullable: true
        parentUnitId:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [id, code, name, isActive]
...